Validate the node existence before path change
This requires explicit usage of !cd to change to not existing path.

Change to invalid path can be handy if there is going to be device
mounted in the future to that location. On the other hand it is only a
corner case when this is handy and thus default change now validates the
node existence.