cli: fix color scheme with white terinal
This just drops our custom style and uses ansi colors directly. That
seems to correctly work while if we pass it trough our style it for some
reason doesn't work correctly to output ansi escape sequences for
primary colors instead of 256 colors.

It is also just easier to specify ansi color name directly in place
where we select the style.