cli: correctly manage the method call timeout
Previously the method call timeouts were not handled and thus resulted
into hidden exception and shvcli got stuck. This instead handles these
timeouts and on top of that adds options to control timeout.

The default timeout is chosen to be shorter than what is the default for
pySHV because here we want to inform user somewhat way sooner than in
the regular code. User can always increase the timeout and attempts if
that is something desired.