cli: Better handle Ctrl-C
The Ctrl-C is propagated in asyncio as CancelledError but that cancels
the main task. This task would have to be uncancelled which is not ideal
as we might not be sure which task should be uncancelled. Thus we
instead do what prompt-toolkit does as well, we set our own signal
handler. This handler cancels the task that handles inserted command
line.