cli: use Ctrl-o to force submission of invalid output
With increased level of validation SHVCLI performs it is sometimes hard
to submit calls to the methods that do not correctly hint on their type.

The Ctrl-o was choosen because for Emacs it functions pretty much same
as enter and in Vi normal mode jump lists are not required in terminal.
There is one TODO for Vi noting that this keybinding should be disabled
in input mode because it replaces temporally escape to the normal mode.